style: view side bar
