+++ apps/frontend/src/lib/calendar/CalendarIndex.svelte
@@ -24,7 +24,7 @@
 		<svelte:component this={map[field.type]} {field} />
 	</div>
 {:else}
-	<div class="flex items-center justify-center h-screen w-full bg-gray-100 dark:bg-slate-800/80">
+	<div class="flex items-center justify-center h-full w-full bg-gray-100 dark:bg-slate-800/80">
 		<Card.Root class="w-96">
 			<Card.Header>
 				<CalendarConfig />
