+++ apps/frontend/src/lib/table/KanbanIndex.svelte
@@ -20,9 +20,11 @@
 </script>
 
 {#if field}
-	<svelte:component this={map[field.type]} {field} />
+	<div class="h-full overflow-hidden">
+		<svelte:component this={map[field.type]} {field} />
+	</div>
 {:else}
-	<div class="flex items-center justify-center h-screen w-full bg-gray-100 dark:bg-slate-800/80">
+	<div class="flex items-center h-full justify-center w-full bg-gray-100 dark:bg-slate-800/80">
 		<Card.Root class="w-96">
 			<Card.Header>
 				<KanbanConfig />
