+++ apps/frontend/src/lib/table/TableNav.svelte
@@ -6,12 +6,30 @@
 	import * as Tooltip from '$lib/components/ui/tooltip'
 	import { updateTableModal } from '$lib/store/modal'
 	import { t } from '$lib/i18n'
+	import { sidebarCollapsed } from '$lib/store/ui'
 
 	const table = getTable()
 </script>
 
-<div class="flex items-center px-3 py-2 gap-3">
-	<p class="text-sm font-medium text-gray-600">
+<div class="flex items-center px-3 py-2 gap-3 border-b relative">
+	{#if $sidebarCollapsed}
+		<div class="ml-2">
+			<Tooltip.Root>
+				<Tooltip.Trigger>
+					<button on:click={() => ($sidebarCollapsed = false)}>
+						<i class="ti ti-layout-sidebar-left-expand text-lg text-gray-500 dark:hover:text-gray-100" />
+					</button>
+				</Tooltip.Trigger>
+				<Tooltip.Content
+					class="px-2 py-1.5 text-xs font-semibold text-gray-800 bg-gray-100 border border-gray-200 rounded-lg dark:bg-gray-600 dark:text-gray-100 dark:border-gray-500"
+				>
+					<kbd> Command + b </kbd>
+				</Tooltip.Content>
+			</Tooltip.Root>
+		</div>
+	{/if}
+
+	<p class="text-sm font-medium text-gray-800">
 		{$table.name.value}
 	</p>
