+++ apps/frontend/src/routes/(authed)/t/[tableId]/[[viewId]]/+layout.svelte
@@ -2,12 +2,10 @@
 	import { currentTable } from '$lib/store/table'
 	import TableNav from '$lib/table/TableNav.svelte'
 	import TableToolBar from '$lib/table/TableToolBar.svelte'
-	import TableViewsTab from '$lib/table/TableViewsTab.svelte'
 </script>
 
 <div class="w-full h-full flex flex-col">
 	<TableNav />
-	<TableViewsTab />
 	<TableToolBar />
 	<slot />
 </div>
